End of Practical 2

+++ Practical 2/ProgrammingExercises/MinutesConversion.java
@@ -5,8 +5,8 @@ public class MinutesConversion {
     {
         String result;
         result = JOptionPane.showInputDialog("Enter a number of minutes");
-        int hours = Integer.parseInt(result)/60;
-        int days = hours/24;
+        double hours = Integer.parseInt(result)/60;
+        double days = hours/24;
         JOptionPane.showMessageDialog(null, result+" minutes equals "+hours+" hours and equals "+days+" days");
     }
 }
